requirements.txt
streamlit
pandas
numpy
scikit-learn
spacy
PyPDF2
python-docx